test(proxy): run key model display tests under proxy/auth for Codecov
Move GET /key/{id}/models helper tests from management_endpoints to
tests/test_litellm/proxy/auth so they execute in the proxy-auth workflow
alongside other model_checks coverage. Expand cases for search, compact
sentinels, ungrouped, and helper edge paths.

Made-with: Cursor
